Added code for a special console window from dhewg.
We use the console window to finally display text. However, we have to move SLOWLY or else we will cause a crash and never get any useful info beyond a black screen. Thus, I've placed an infinite loop to demonstrate that output is fixed (but again, many bugs not perfected)

Please see the source code for more details.